% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_gpx}
\alias{read_gpx}
\title{Read a .gpx file into a data.frame}
\usage{
read_gpx(file)
}
\arguments{
\item{file}{A path to a .gpx file}
}
\value{
List of data frames.
}
\description{
Read a .gpx file into a list of data.frames
}
\examples{
\dontrun{
hikes = read_gpx('hiking_file.gpx')
hikes$tracks
hikes$routes
}
}
